+++ package.json
@@ -46,6 +46,7 @@
   },
   "devDependencies": {
     "angular2-template-loader": "^0.5.0",
+    "autoprefixer": "^6.4.1",
     "awesome-typescript-loader": "^1.1.1",
     "browser-sync": "^2.13.0",
     "browser-sync-webpack-plugin": "^1.0.3",
@@ -61,6 +62,8 @@
     "karma-htmlfile-reporter": "^0.2.2",
     "karma-jasmine": "^0.3.8",
     "node-sass": "^3.9.0",
+    "postcss-loader": "^0.13.0",
+    "precss": "^1.4.0",
     "protractor": "^3.3.0",
     "raw-loader": "^0.5.1",
     "rimraf": "^2.5.2",
