+++ src/app/home/monthly/monthly.component.ts
@@ -1,20 +1,17 @@
-import {Component} from '@angular/core';
-import {Header} from "../../header/header";
-import {Footer} from "../../footer/footer";
+import {Component} from "@angular/core";
+import {Calendar} from "../calendar";
 import {CalendarService} from "../../shared/calendar/calendar.service";
-import {Calendar} from "../calendar/calendar.component";
+
 
 const template = require('./monthly.html');
+
 @Component({
     selector: 'monthly-view',
-    template: template,
-    providers: [
-        Header,
-        Footer,
-        Calendar,
-        CalendarService
-    ]
-
+    providers: [Calendar, CalendarService],
+    template: template
 })
 export class MonthlyViewComponent {
+
+    constructor() {
+    }
 }
