+++ src/app/home/yearly/yearly.component.ts
@@ -1,17 +1,20 @@
-import {Component} from "@angular/core";
-import {Calendar} from "../calendar";
+import {Component} from '@angular/core';
+import {Header} from "../../header/header";
+import {Footer} from "../../footer/footer";
 import {CalendarService} from "../../shared/calendar/calendar.service";
-
+import {Calendar} from "../calendar/calendar.component";
 
 const template = require('./yearly.html');
-
 @Component({
     selector: 'yearly-view',
-    providers: [Calendar, CalendarService],
-    template: template
+    template: template,
+    providers: [
+        Header,
+        Footer,
+        Calendar,
+        CalendarService
+    ]
+
 })
 export class YearlyViewComponent {
-
-    constructor() {
-    }
 }
