+++ src/app/login/login.component.ts
@@ -25,7 +25,7 @@ export class Login {
             response => {
                 localStorage.setItem('id_token', response.json().token_key);
                 localStorage.setItem('user', JSON.stringify(response.json()));
-                this.router.navigate(['/home/monthly']);
+                this.router.navigate(['/home/yearly']);
             },
             error => {
                 alert(error.text());
